+++ src/transport/drd_rdp_listener.h
@@ -11,6 +11,11 @@ G_BEGIN_DECLS
 #define DRD_TYPE_RDP_LISTENER (drd_rdp_listener_get_type())
 G_DECLARE_FINAL_TYPE(DrdRdpListener, drd_rdp_listener, DRD, RDP_LISTENER, GSocketService)
 
+typedef gboolean (*DrdRdpListenerDelegateFunc)(DrdRdpListener *listener,
+                                               GSocketConnection *connection,
+                                               gpointer user_data,
+                                               GError **error);
+
 DrdRdpListener *drd_rdp_listener_new(const gchar *bind_address,
                                      guint16 port,
                                      DrdServerRuntime *runtime,
@@ -23,5 +28,11 @@ DrdRdpListener *drd_rdp_listener_new(const gchar *bind_address,
 gboolean drd_rdp_listener_start(DrdRdpListener *self, GError **error);
 void drd_rdp_listener_stop(DrdRdpListener *self);
 DrdServerRuntime *drd_rdp_listener_get_runtime(DrdRdpListener *self);
+void drd_rdp_listener_set_delegate(DrdRdpListener *self,
+                                   DrdRdpListenerDelegateFunc func,
+                                   gpointer user_data);
+gboolean drd_rdp_listener_adopt_connection(DrdRdpListener *self,
+                                           GSocketConnection *connection,
+                                           GError **error);
 
 G_END_DECLS